requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Scheduling
schedule>=1.2.0

# Twitter/X API
tweepy>=4.14.0

# Discord Bot
discord.py>=2.3.0
aiohttp>=3.9.0

# Optional: for webhook fallback
# Already included via requests
